Move enums to core module
